requirements.txt
# Core Dependencies
streamlit==1.28.1
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.7
python-dotenv==1.0.0

# Vector Database
pinecone-client==2.2.4
chromadb==0.4.18
faiss-cpu==1.7.4

# Web Tools
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
google-search-results==2.4.2

# Data Processing
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
plotly==5.17.0

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Security
cryptography==41.0.8
bcrypt==4.1.1
